Moving 2 new tests to jetstream test file
